PS2 Mouse: Switched tp sensitivity setting from float to int

+++ include/zmk/mouse_ps2.h
@@ -9,8 +9,8 @@
 int zmk_ps2_set_sampling_rate_decr();
 int zmk_ps2_set_sampling_rate_incr();
 
-int zmk_ps2_tp_sensitivity_incr(float incr_amount);
-int zmk_ps2_tp_sensitivity_decr(float decr_amount);
+int zmk_ps2_tp_sensitivity_incr(int incr_amount);
+int zmk_ps2_tp_sensitivity_decr(int decr_amount);
 
 int zmk_ps2_tp_neg_inertia_incr(int incr_amount);
 int zmk_ps2_tp_neg_inertia_decr(int decr_amount);
